Refactor Game Over to use scene transition instead of overlay
- Moved game over detection from HealthComponent to Map.removeEntity()
- Simplified MapDelegate to just gameOver() method
- Created dedicated GameOverScene instead of overlay
- GameScene now transitions to GameOverScene with fade effect
- Cleaner separation of concerns and better architecture
